+++ flang/lib/Lower/OpenACC.cpp
@@ -2531,13 +2531,13 @@ genDeclareInFunction(Fortran::lower::AbstractConverter &converter,
       genDataOperandOperations<mlir::acc::DevicePtrOp>(
           devicePtrClause->v, converter, semanticsContext, stmtCtx,
           dataClauseOperands, mlir::acc::DataClause::acc_deviceptr,
-          /*structured=*/true, /*implicit=*/false);
+          /*structured=*/true, /*implicit=*/false, /*setDeclareAttr=*/true);
     } else if (const auto *linkClause =
                    std::get_if<Fortran::parser::AccClause::Link>(&clause.u)) {
       genDataOperandOperations<mlir::acc::DeclareLinkOp>(
           linkClause->v, converter, semanticsContext, stmtCtx,
           dataClauseOperands, mlir::acc::DataClause::acc_declare_link,
-          /*structured=*/true, /*implicit=*/false);
+          /*structured=*/true, /*implicit=*/false, /*setDeclareAttr=*/true);
     } else if (const auto *deviceResidentClause =
                    std::get_if<Fortran::parser::AccClause::DeviceResident>(
                        &clause.u)) {
@@ -2546,7 +2546,7 @@ genDeclareInFunction(Fortran::lower::AbstractConverter &converter,
           deviceResidentClause->v, converter, semanticsContext, stmtCtx,
           dataClauseOperands,
           mlir::acc::DataClause::acc_declare_device_resident,
-          /*structured=*/true, /*implicit=*/false);
+          /*structured=*/true, /*implicit=*/false, /*setDeclareAttr=*/true);
       deviceResidentEntryOperands.append(
           dataClauseOperands.begin() + crtDataStart, dataClauseOperands.end());
     } else {
@@ -2556,27 +2556,24 @@ genDeclareInFunction(Fortran::lower::AbstractConverter &converter,
   }
   builder.create<mlir::acc::DeclareEnterOp>(loc, dataClauseOperands);
 
-  if (!createEntryOperands.empty() || !copyEntryOperands.empty() ||
-      !copyoutEntryOperands.empty() || !deviceResidentEntryOperands.empty()) {
-    // Attach declare exit operation generation to function context.
-    fctCtx.attachCleanup([&builder, loc, dataClauseOperands,
-                          createEntryOperands, copyEntryOperands,
-                          copyoutEntryOperands, deviceResidentEntryOperands]() {
-      builder.create<mlir::acc::DeclareExitOp>(loc, dataClauseOperands);
-      genDataExitOperations<mlir::acc::CreateOp, mlir::acc::DeleteOp>(
-          builder, createEntryOperands, /*structured=*/true,
-          /*implicit=*/false);
-      genDataExitOperations<mlir::acc::DeclareDeviceResidentOp,
-                            mlir::acc::DeleteOp>(
-          builder, deviceResidentEntryOperands, /*structured=*/true,
-          /*implicit=*/false);
-      genDataExitOperations<mlir::acc::CopyinOp, mlir::acc::CopyoutOp>(
-          builder, copyEntryOperands, /*structured=*/true, /*implicit=*/false);
-      genDataExitOperations<mlir::acc::CreateOp, mlir::acc::CopyoutOp>(
-          builder, copyoutEntryOperands, /*structured=*/true,
-          /*implicit=*/false);
-    });
-  }
+  // Attach declare exit operation generation to function context.
+  fctCtx.attachCleanup([&builder, loc, dataClauseOperands, createEntryOperands,
+                        copyEntryOperands, copyoutEntryOperands,
+                        deviceResidentEntryOperands]() {
+    builder.create<mlir::acc::DeclareExitOp>(loc, dataClauseOperands);
+    genDataExitOperations<mlir::acc::CreateOp, mlir::acc::DeleteOp>(
+        builder, createEntryOperands, /*structured=*/true,
+        /*implicit=*/false);
+    genDataExitOperations<mlir::acc::DeclareDeviceResidentOp,
+                          mlir::acc::DeleteOp>(
+        builder, deviceResidentEntryOperands, /*structured=*/true,
+        /*implicit=*/false);
+    genDataExitOperations<mlir::acc::CopyinOp, mlir::acc::CopyoutOp>(
+        builder, copyEntryOperands, /*structured=*/true, /*implicit=*/false);
+    genDataExitOperations<mlir::acc::CreateOp, mlir::acc::CopyoutOp>(
+        builder, copyoutEntryOperands, /*structured=*/true,
+        /*implicit=*/false);
+  });
 }
 
 static void